fix(lint): clear all golangci-lint v2.11 findings
- .golangci.yml: errcheck excludes for fmt.Fprint*/tabwriter.Flush/
  io.Closer.Close (always-nil writers); drop revive's
  redefines-builtin-id (Go 1.21+ added max/min) and exported
  rules; widen the test-file exclusion to also skip unparam.
- defer store.Close() → defer func() { _ = store.Close() }() in the
  three CLI/MCP entry points so the explicit-discard reads cleanly.
- internal/optimizer/dedupe: drop the now-unused messageProbe stub
  and its compile-time sink.
- internal/otlp/exporter.go: remove redundant int64() conversion.
- internal/cli/spend.go: writeSpendText now propagates tw.Flush()
  errors so it's not always-nil.
- internal/rbac/rbac_test.go: rewrite the defensive-copy test so
  the append result is the dead store (the side effect is the
  point) rather than a re-assignment.
- Drop two stale //nolint:unparam directives now that the
  test-file exclusion covers them centrally.

+++ internal/otlp/exporter.go
@@ -250,7 +250,7 @@ func workflowAttributes(p *eventschema.WorkflowEvent) []kv {
 	out := []kv{
 		stringKV(eventschema.AttrTokenOpsWorkflowID, p.WorkflowID),
 		stringKV(eventschema.AttrTokenOpsWorkflowState, string(p.State)),
-		intKV(eventschema.AttrTokenOpsWorkflowStepCount, int64(p.StepCount)),
+		intKV(eventschema.AttrTokenOpsWorkflowStepCount, p.StepCount),
 	}
 	if p.AgentID != "" {
 		out = append(out, stringKV(eventschema.AttrTokenOpsAgentID, p.AgentID))
